+++ src/roq/adapter/clickhouse/modify_order.hpp
@@ -31,7 +31,7 @@ struct ModifyOrder final {
 
  private:
   Column<std::string_view, true> account_;
-  Column<uint32_t> order_id_;
+  Column<uint64_t> order_id_;
   Column<std::string_view, true> request_template_;
   Column<double> quantity_;
   Column<double> price_;
