+++ src/roq/adapter/clickhouse/trade.hpp
@@ -32,7 +32,7 @@ struct Trade final {
  private:
   Column<uint16_t> stream_id_;
   Column<std::string_view, true> account_;
-  Column<uint32_t> order_id_;
+  Column<uint64_t> order_id_;
   Column<std::string_view, true> exchange_;
   Column<std::string_view, true> symbol_;
   Column<Side, true> side_;
